Add linux arm64 support to build.rs
A series of coincidences meant that I noticed that this arch had
builds, and seems to be kept relatively up to date.

"Tested" in that it builds on my asahi linux machine.
I wonder how valve tests it.
monado + opencomposite or xrizer on an arm64 machine would probably
do the trick, though.